Added submitting the payment

+++ ClientApp/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -74,12 +74,25 @@ export class CheckoutPaymentComponent implements OnInit {
 
     const orderToCreate = this.getOrderToCreate(basket);
     if (!orderToCreate) return;
+
     this.checkoutService.createOrder(orderToCreate).subscribe({
       next: (order) => {
         this.toastr.success('Order created successfully');
-        this.basketService.deleteLocalBasket();
-        const navigationExtras: NavigationExtras = { state: order };
-        this.router.navigate(['checkout/success'], navigationExtras);
+        this.stripe?.confirmCardPayment(basket.clientSecret!, {
+          payment_method: {
+            card: this.cardNumber!,
+            billing_details: {
+              name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+            }
+          }
+        }).then(result => {
+          console.log(result);
+          if (result.paymentIntent) {
+            this.basketService.deleteLocalBasket();
+            const navigationExtras: NavigationExtras = { state: order };
+            this.router.navigate(['checkout/success'], navigationExtras);
+          }
+        });
       },
     });
   }
